✨[feat] : 영수증을 추가할 때 총합을 여행총합에서 빼기

+++ src/main/java/com/wevel/wevel_server/tripInfo/dto/TripInfoDTO.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class TripInfoDTO {
     private String tripName;
     private String country;
-    private Long totalBudget;
+    private double totalBudget;
     private Date startDate;
     private Date endDate;
